+++ client/src/components/MainPage/mainComponents/Table/index.js
@@ -1,0 +1,1 @@
+export { default } from './Table'
