+++ client/src/components/Navbar/Navbar.js
@@ -4,14 +4,57 @@ import { withStyles } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
-import IconButton from '@material-ui/core/IconButton';
 import MenuIcon from '@material-ui/icons/Menu';
-import LeftDrawer from './navbarComponents/LeftDrawer.js'
+
+import SwipeableDrawer from '@material-ui/core/SwipeableDrawer';
+import List from '@material-ui/core/List';
+import Divider from '@material-ui/core/Divider';
+import ListItem from '@material-ui/core/ListItem';
+import ListItemIcon from '@material-ui/core/ListItemIcon';
+import ListItemText from '@material-ui/core/ListItemText';
+import CalendarTodayIcon from '@material-ui/icons/CalendarToday';
+import PieChartIcon from '@material-ui/icons/PieChart';
+import AttachMoneyIcon from '@material-ui/icons/AttachMoney';
+import Grid from '@material-ui/core/Grid';
+import AccountCircle from '@material-ui/icons/AccountCircle';
 
 
-const styles = {
+import Button from '@material-ui/core/Button';
+import ClickAwayListener from '@material-ui/core/ClickAwayListener';
+import Grow from '@material-ui/core/Grow';
+import Paper from '@material-ui/core/Paper';
+import Popper from '@material-ui/core/Popper';
+import MenuItem from '@material-ui/core/MenuItem';
+import MenuList from '@material-ui/core/MenuList';
+import Avatar from '@material-ui/core/Avatar';
+import { Link } from 'react-router-dom';
+
+
+import Tabs from '@material-ui/core/Tabs';
+import NoSsr from '@material-ui/core/NoSsr';
+import Tab from '@material-ui/core/Tab';
+
+
+function TabContainer(props) {
+  return (
+    <Typography component="div" style={{ padding: 8 * 3 }}>
+      {props.children}
+    </Typography>
+  );
+}
+TabContainer.propTypes = {
+  children: PropTypes.node.isRequired,
+};
+
+function LinkTab(props) {
+  return <Tab component="a" onClick={event => event.preventDefault()} {...props} />;
+}
+
+const styles = theme => ({
   root: {
     flexGrow: 1,
+    display: 'flex',
+    backgroundColor: theme.palette.background.paper,
   },
   grow: {
     flexGrow: 1,
@@ -20,28 +63,148 @@ const styles = {
     marginLeft: -12,
     marginRight: 20,
   },
-};
+  list: {
+    width: 200,
+  },
 
-function ButtonAppBar(props) {
-  const { classes } = props;
+  fullList: {
+    width: 'auto',
+  },
+  toolbar: {
+    alignItems: 'center',
+    justifyContent: 'space-between',
+  },
+  bigAvatar: {
+    width: 200,
+    height: 150,
+  },
+  paper: {
+    marginRight: theme.spacing.unit * 2,
+  },
+});
+
+
+class NavBar extends React.Component {
+  state = {
+    left: false,
+    open: false,
+    value: 0,
+  };
+  
+  handleChange = (event, value) => {
+    this.setState({ value });
+  };
+
+  toggleDrawer = (side, open) => () => {
+    this.setState({
+      [side]: open,
+    });
+  };
+  handleToggle = () => {
+    this.setState(state => ({ open: !state.open }));
+  };
+
+  handleClose = event => {
+    if (this.anchorEl.contains(event.target)) {
+      return;
+    }
+
+    this.setState({ open: false });
+  };
+
+render () {
+
+  const { classes } = this.props;
+  const { open } = this.state;
+  const { value } = this.state;
+
+    const sideList = (
+      <div className={classes.list}>
+        <List>
+        <Grid container justify="center" alignItems="center"> 
+          <img src="./images/piggy.jpg" alt='piggy'/>
+        </Grid>
+        </List>
+        <Divider /> 
+        <List>
+        {['Calendar', 'Chart', 'Budgets','About us'].map((text, index) => (
+            <ListItem button key={text}>
+              <ListItemIcon>{index  === 0 ? <CalendarTodayIcon /> : index  === 1 ? <PieChartIcon /> : 
+              index === 2 ? <AttachMoneyIcon />  : index === 3 ?<AccountCircle /> : <AccountCircle />}</ListItemIcon>
+              <ListItemText primary={text} />
+            </ListItem>
+            
+          ))}
+        </List>
+      </div>
+    );
   return (
+    <>
     <div className={classes.root}>
       <AppBar position="static">
-        <Toolbar>
-          <IconButton className={classes.menuButton} color="inherit" aria-label="Menu">
-            <MenuIcon />
-          </IconButton>
-          <Typography variant="h6" color="inherit" className={classes.grow}>
-          <LeftDrawer/>
+          <div>
+        <AppBar position="fixed" color="primary" className={classes.appBar}>
+        <Toolbar className={classes.toolbar}>      
+          <Button onClick={this.toggleDrawer('left', true)}><MenuIcon /></Button>  
+          <Typography>
+          Piggy Bank
           </Typography>
+          <Button
+            buttonRef={node => {
+              this.anchorEl = node;
+            }}
+            aria-owns={open ? 'menu-list-grow' : undefined}
+            aria-haspopup="true"
+            onClick={this.handleToggle}
+          >
+            <Avatar alt="Remy Sharp" src="./images/taylor.jpg"/>
+          </Button>
+          <Popper open={open} anchorEl={this.anchorEl} transition disablePortal>
+            {({ TransitionProps, placement }) => (
+              <Grow
+                {...TransitionProps}
+                id="menu-list-grow"
+                style={{ transformOrigin: placement === 'bottom' ? 'center top' : 'center bottom' }}
+              >
+                <Paper>
+                  <ClickAwayListener onClickAway={this.handleClose}>
+                    <MenuList>
+                      <MenuItem><Link to='/AccountInfo'>My Account</Link></MenuItem>
+                      <MenuItem><Link to='/'>Log Out</Link></MenuItem>
+                    </MenuList>
+                  </ClickAwayListener>
+                </Paper>
+              </Grow>
+            )}
+          </Popper>
         </Toolbar>
       </AppBar>
+        <SwipeableDrawer
+          open={this.state.left}
+          onClose={this.toggleDrawer('left', false)}
+          onOpen={this.toggleDrawer('left', true)}
+        >
+          <div
+            tabIndex={0}
+            role="button"
+          x
+            onKeyDown={this.toggleDrawer('left', false)}
+          >
+            {sideList}
+          </div>
+          
+        </SwipeableDrawer>   
+      </div>
+      </AppBar>
     </div>
+
+</>
   );
 }
+}
 
-ButtonAppBar.propTypes = {
+NavBar.propTypes = {
   classes: PropTypes.object.isRequired,
 };
 
-export default withStyles(styles)(ButtonAppBar);
+export default withStyles(styles)(NavBar);
